requirements.txt
# Alpha Prophet CLI Requirements
anthropic>=0.39.0
python-dotenv>=1.0.0
pandas>=2.0.0
openpyxl>=3.1.0
